extracted FeedStoreSpecs helpers methods to a shared XCTestCase extension to be used by another FeedStore test implementations
